Update code to use errno values (#44)
This updates all code that throws FileSystemException to properly set
the corresponding osError.errorCode, which allows us to update the
tests to expect the error code rather than the error message. This in turn
allows the tests to be more robust across locales and platforms.

To deal with the fact that different platforms have different errno values
(which will be exposed and tested via LocalFileSystem), this change
creates an ErrorCodes class with static getters for the set of errno
values that are common to all platforms. This code can be removed once
dart-lang/sdk#28860 is resolved and live.

Fixes #43