Re-organize repo
Before, package:file/testing.dart was a library containing testing
utilities.  That library depended on package:test.  However, we
explicitly didn't declare a dependency on `test` in the pubspec.yaml
because most users of package:file don't want that dependency.

Unfortunatley, there's no way to have library-specific sections within
a pubspec.yaml.  So in order to declare that those testing utilities
did in fact depend on `package:test`, we had to publish them as a
separate package.  This change does just that, as `package:file_testing`.