Switch to new-style mixins so that super.noSuchMethod can be called. (#117)
The ability for a class used as a mixin to call super has been removed
from Dart, but due to an analyzer bug some usages slipped through the
cracks.  The analyzer bug has now been fixed, so we need a
corresponding fix to package:file.

The solution is to change the mixin classes to use the new "mixin"
syntax.